+++ src/commands/monitors/run.rs
@@ -7,7 +7,7 @@ use anyhow::Result;
 use clap::{Arg, ArgAction, ArgMatches, Command};
 use console::style;
 
-use sentry::protocol::{MonitorCheckIn, MonitorCheckInStatus};
+use sentry::protocol::{MonitorCheckIn, MonitorCheckInStatus, MonitorConfig, MonitorSchedule};
 use sentry::types::Dsn;
 
 use crate::api::{Api, ApiCreateMonitorCheckIn, ApiMonitorCheckInStatus, ApiUpdateMonitorCheckIn};
@@ -45,6 +45,40 @@ pub fn make_command(command: Command) -> Command {
                 .num_args(1..)
                 .last(true),
         )
+        .arg(Arg::new("schedule").short('s').long("schedule").help(
+            "Configure the cron monitor with the given schedule (crontab format). \
+             Enclose the schedule in quotes to ensure your command line environment \
+             parses the argument correctly.",
+        ))
+        .arg(
+            Arg::new("checkin_margin")
+                .long("check-in-margin")
+                .value_parser(clap::value_parser!(u64).range(1..))
+                .requires("schedule")
+                .help(
+                    "The allowed margin of minutes after the expected check-in time that the \
+                     monitor will not be considered missed for. Requires --schedule.",
+                ),
+        )
+        .arg(
+            Arg::new("max_runtime")
+                .long("max-runtime")
+                .value_parser(clap::value_parser!(u64).range(1..))
+                .requires("schedule")
+                .help(
+                    "The allowed duration in minutes that the monitor may be in progress for \
+                     before being considered failed due to timeout. Requires --schedule.",
+                ),
+        )
+        .arg(
+            Arg::new("timezone")
+                .long("timezone")
+                .requires("schedule")
+                .help(
+                    "A tz database string (e.g. \"Europe/Vienna\") representing the monitor's \
+             execution schedule's timezone. Requires --schedule.",
+                ),
+        )
 }
 
 fn run_program(args: Vec<&String>, monitor_slug: &str) -> (bool, Option<i32>, Duration) {
@@ -75,6 +109,7 @@ fn dsn_execute(
     args: Vec<&String>,
     monitor_slug: &str,
     environment: &str,
+    monitor_config: Option<MonitorConfig>,
 ) -> (bool, Option<i32>) {
     let check_in_id = Uuid::new_v4();
 
@@ -84,7 +119,7 @@ fn dsn_execute(
         status: MonitorCheckInStatus::InProgress,
         duration: None,
         environment: Some(environment.to_string()),
-        monitor_config: None,
+        monitor_config,
     };
 
     with_sentry_client(dsn.clone(), |c| c.send_envelope(open_checkin.into()));
@@ -158,6 +193,19 @@ fn token_execute(
     (success, code, None)
 }
 
+fn parse_monitor_config_args(matches: &ArgMatches) -> Result<Option<MonitorConfig>> {
+    let Some(schedule) = matches.get_one::<String>("schedule") else {
+        return Ok(None);
+    };
+    let schedule = MonitorSchedule::from_crontab(schedule)?;
+    Ok(Some(MonitorConfig {
+        schedule,
+        checkin_margin: matches.get_one("checkin_margin").copied(),
+        max_runtime: matches.get_one("max_runtime").copied(),
+        timezone: matches.get_one("timezone").cloned(),
+    }))
+}
+
 pub fn execute(matches: &ArgMatches) -> Result<()> {
     let config = Config::current();
     let dsn = config.get_dsn().ok();
@@ -172,13 +220,18 @@ pub fn execute(matches: &ArgMatches) -> Result<()> {
     let args: Vec<_> = matches.get_many::<String>("args").unwrap().collect();
     let monitor_slug = matches.get_one::<String>("monitor_slug").unwrap();
     let environment = matches.get_one::<String>("environment").unwrap();
+    let monitor_config = parse_monitor_config_args(matches)?;
 
     let (success, code) = match dsn {
         // Use envelope API when dsn is provided. This is the prefered way to create check-ins,
         // and the legacy API will be removed in the next major CLI version.
-        Some(dsn) => dsn_execute(dsn, args, monitor_slug, environment),
+        Some(dsn) => dsn_execute(dsn, args, monitor_slug, environment, monitor_config),
         // Use legacy API when DSN is not provided
         None => {
+            if monitor_config.is_some() {
+                anyhow::bail!("Crons monitor upserts are only supported with DSN auth. Please try again with \
+                               DSN auth or repeat the command without the `schedule` argument.");
+            }
             let (success, code, err) = token_execute(args, monitor_slug, environment);
             if let Some(e) = err {
                 if matches.get_flag("allow_failure") {